requirements.txt
Django==4.2.8
djangorestframework==3.14.0
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
djoser==2.2.1
djangorestframework-simplejwt==5.5.1
Pillow==10.1.0
django-filter==23.5
gunicorn==21.2.0
whitenoise==6.6.0
django-cloudinary-storage==0.3.0
cloudinary==1.36.0
dj-database-url==2.1.0